requirements.txt
flask
pymupdf
python-docx
matplotlib
pandas
openai
reportlab
markdown